+++ apps/miranum-jsonforms/src/adapter/out/workspace.ts
@@ -1,11 +1,9 @@
 import { Uri, workspace } from "vscode";
-import { singleton } from "tsyringe";
 import {
     CreateFileOutPort,
     FormPreviewSettingsOutPort,
 } from "../../application/ports/out";
 
-@singleton()
 export class VsCodeFormPreviewSettingsAdapter implements FormPreviewSettingsOutPort {
     getRenderer(): string {
         const setting = workspace
@@ -20,7 +18,6 @@ export class VsCodeFormPreviewSettingsAdapter implements FormPreviewSettingsOutP
     }
 }
 
-@singleton()
 export class VsCodeFileSystemAdapter implements CreateFileOutPort {
     private readonly fs = workspace.fs;